feat: SectionRepositorty 구조 수정
'SectionRepository'의 구간을 관리하는 변수가 잘못 구성되어 있었음.
map에서 -> Section 클래스로 변경함.

그에 따른 영향받는 다른 도메인 코드 수정함.

모든 'repositry' 클래스는 최소한의 CRUD 로직만 포함하도록 다시 구성함.

+++ src/main/java/subway/service/LineService.java
@@ -1,23 +1,35 @@
 package subway.service;
 
-import subway.domain.Line;
-import subway.domain.LineRepository;
-import subway.domain.SectionRepository;
-import subway.domain.Station;
+import subway.domain.*;
 
 import java.util.List;
 
 public class LineService {
+    private static final int LINE_NAME_MIN_SIZE = 2;
+    private static final String LINE_NAME_TOO_SHORT_ERROR = "노선 이름은 " + LINE_NAME_MIN_SIZE + "글자 이상입니다.";
+    private static final String LINE_ALREADY_EXISTS_ERROR = "이미 등록된 노선 이름입니다.";
+    private static final String LINE_DELETE_ERROR = "삭제되지 않았습니다.";
+    private static final String LINE_REGISTER_ERROR = "노선 등록 실패";
+    private static final String STATION_NOT_FOUND_ERROR = "해당 역이 존재하지 않습니다.";
 
     public static void addLine(Line line, Station upStation, Station downStation) {
-        // line 추가
+        if (line.getName().length() < LINE_NAME_MIN_SIZE) {
+            throw new IllegalArgumentException(LINE_NAME_TOO_SHORT_ERROR);
+        }
+
+        if (LineRepository.lines().contains(line)) {
+            throw new IllegalArgumentException(LINE_ALREADY_EXISTS_ERROR);
+        }
+
         LineRepository.addLine(line);
 
-        // section(구간) 초기값 설정
         initSection(line, downStation, upStation);
     }
 
     public static void deleteLineByName(String deleteLineName) {
+        if (!LineRepository.deleteLineByName(deleteLineName)) {
+            throw new IllegalArgumentException(deleteLineName + " " + LINE_DELETE_ERROR);
+        }
         LineRepository.deleteLineByName(deleteLineName);
     }
 
@@ -27,9 +39,16 @@ public class LineService {
 
     private static void initSection(Line line, Station downStation, Station upStation) {
         try {
-            SectionRepository.initSection(line, downStation, upStation);
+            Station findDownStation = StationRepository.findById(downStation)
+                    .orElseThrow(() -> new IllegalArgumentException(LINE_REGISTER_ERROR + " - " + STATION_NOT_FOUND_ERROR));
+
+            Station findUpStation = StationRepository.findById(upStation)
+                    .orElseThrow(() -> new IllegalArgumentException(LINE_REGISTER_ERROR + " - " + STATION_NOT_FOUND_ERROR));
+
+            Section section = Section.of(line, findDownStation, findUpStation);
+            SectionRepository.addSection(section);
         } catch (Exception e) {
-            LineRepository.deleteLineByName(line.getName());
+            deleteLineByName(line.getName());
             throw e;
         }
     }
